+++ libfreeipmi/include/freeipmi/util/ipmi-lan-util.h
@@ -26,35 +26,35 @@ extern "C" {
 #include <stdint.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_lan_check_session_sequence_number (fiid_obj_t obj_lan_session_hdr,
-                                               uint32_t session_sequence_number);
+int ipmi_lan_check_session_sequence_number (fiid_obj_t obj_lan_session_hdr,
+                                            uint32_t session_sequence_number);
 
-int8_t ipmi_lan_check_session_id (fiid_obj_t obj_lan_session_hdr,
-                                  uint32_t session_id);
+int ipmi_lan_check_session_id (fiid_obj_t obj_lan_session_hdr,
+                               uint32_t session_id);
 
-int8_t ipmi_lan_check_session_authentication_code (fiid_obj_t obj_lan_session_hdr_rs,
-                                                   fiid_obj_t obj_lan_msg_hdr_rs,
-                                                   fiid_obj_t obj_cmd,
-                                                   fiid_obj_t obj_lan_msg_trlr_rs,
-                                                   uint8_t authentication_type,
-                                                   uint8_t *authentication_code_data,
-                                                   uint32_t authentication_code_data_len);
+int ipmi_lan_check_session_authentication_code (fiid_obj_t obj_lan_session_hdr_rs,
+                                                fiid_obj_t obj_lan_msg_hdr_rs,
+                                                fiid_obj_t obj_cmd,
+                                                fiid_obj_t obj_lan_msg_trlr_rs,
+                                                uint8_t authentication_type,
+                                                uint8_t *authentication_code_data,
+                                                uint32_t authentication_code_data_len);
 
-int8_t ipmi_lan_check_packet_session_authentication_code (uint8_t *pkt,
-                                                          uint64_t pkt_len,
-                                                          uint8_t authentication_type,
-                                                          uint8_t *authentication_code_data,
-                                                          uint32_t authentication_code_data_len);
+int ipmi_lan_check_packet_session_authentication_code (uint8_t *pkt,
+                                                       uint64_t pkt_len,
+                                                       uint8_t authentication_type,
+                                                       uint8_t *authentication_code_data,
+                                                       uint32_t authentication_code_data_len);
 
-int8_t ipmi_lan_check_net_fn (fiid_obj_t obj_lan_msg_hdr, uint8_t net_fn);
+int ipmi_lan_check_net_fn (fiid_obj_t obj_lan_msg_hdr, uint8_t net_fn);
 
-int8_t ipmi_lan_check_rq_seq (fiid_obj_t obj_lan_msg_hdr, uint8_t rq_seq);
+int ipmi_lan_check_rq_seq (fiid_obj_t obj_lan_msg_hdr, uint8_t rq_seq);
 
-int8_t ipmi_lan_check_checksum (fiid_obj_t obj_lan_msg_hdr,
-                                fiid_obj_t obj_cmd,
-                                fiid_obj_t obj_lan_msg_trlr);
+int ipmi_lan_check_checksum (fiid_obj_t obj_lan_msg_hdr,
+                             fiid_obj_t obj_cmd,
+                             fiid_obj_t obj_lan_msg_trlr);
 
-int8_t ipmi_lan_check_packet_checksum (uint8_t *pkt, uint64_t pkt_len);
+int ipmi_lan_check_packet_checksum (uint8_t *pkt, uint64_t pkt_len);
 
 #ifdef __cplusplus
 }
